requirements.txt
aiofiles==23.2.1
aiohttp==3.8.4
aioredis==2.0.1
aiosignal==1.3.1
annotated-types==0.6.0
anthropic==0.8.1
anyio==4.2.0
anytree==2.12.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.2
cachetools==5.3.2
camel-converter==3.1.1
certifi==2023.11.17
channels==4.0.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses==0.6
dataclasses-json==0.6.3
DBUtils==3.0.3
decorator==5.1.1
deprecation==2.1.0
distro==1.9.0
Django==5.0.1
dnspython==2.4.2
et-xmlfile==1.1.0
exceptiongroup==1.2.0
faiss-cpu==1.7.4
filelock==3.13.1
fire==0.4.0
frozenlist==1.4.1
fsspec==2023.12.2
gitdb==4.0.11
gitignore-parser==0.1.9
GitPython==3.1.40
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.26.1
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.0
grpcio-status==1.60.0
grpcio-tools==1.60.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.2
hyperframe==6.0.1
idna==3.6
isodate==0.6.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-path==0.3.2
jsonschema-spec==0.2.4
jsonschema-specifications==2023.7.1
lancedb==0.4.0
langchain==0.0.352
langchain-community==0.0.10
langchain-core==0.1.8
langsmith==0.0.78
lazy-object-proxy==1.10.0
libcst==1.0.1
loguru==0.6.0
lxml==5.1.0
MarkupSafe==2.1.3
marshmallow==3.20.1
meilisearch==0.21.0
more-itertools==10.2.0
motor==3.3.2
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.24.3
openai==1.6.0
openapi-core==0.18.2
openapi-schema-validator==0.6.2
openapi-spec-validator==0.7.1
openpyxl==3.1.2
overrides==7.4.0
packaging==23.2
pandas==2.0.3
parse==1.20.0
pathable==0.4.3
portalocker==2.8.2
prance==23.6.21.0
proto-plus==1.23.0
protobuf==4.25.1
py==1.11.0
pyahocorasick==2.0.0
pyarrow==14.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
pylance==0.9.0
pymongo==4.6.1
PyMySQL==1.1.0
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
qdrant-client==1.7.0
ratelimiter==1.2.0.post0
redis==5.0.1
referencing==0.30.2
regex==2023.12.25
requests==2.31.0
retry==0.9.2
rfc3339-validator==0.1.4
rpds-py==0.16.2
rsa==4.9
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
semantic-kernel==0.4.3.dev0
semver==3.0.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
socksio==1.0.0
soupsieve==2.5
SQLAlchemy==2.0.25
sqlparse==0.4.4
ta==0.10.2
tenacity==8.2.2
termcolor==2.4.0
tiktoken==0.5.2
tokenizers==0.15.0
tqdm==4.65.0
typer==0.9.0
typing-inspect==0.8.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==1.26.18
websocket-client==1.6.2
websockets==12.0
Werkzeug==3.0.1
win32-setctime==1.1.0
wrapt==1.15.0
yarl==1.9.4
zhipuai==1.0.7
